game/requirements.txt
discord.py==0.16.12
python-dotenv==0.1.0
requests==2.21.0